pyproject.toml
[tool.poetry]
name = "jobs-agent-python"
version = "0.1.0"
description = ""
authors = ["SALHI Mohamed <salhi.mohamed.sd@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.118.0,<0.119.0"
uvicorn = {extras = ["standard"], version = ">=0.37.0,<0.38.0"}
playwright = ">=1.55.0,<2.0.0"
pydantic = ">=2.11.9,<3.0.0"
pydantic-settings = ">=2.11.0,<3.0.0"
structlog = ">=25.4.0,<26.0.0"
python-dateutil = ">=2.9.0.post0,<3.0.0"
aiosmtplib = ">=4.0.2,<5.0.0"
jinja2 = ">=3.1.6,<4.0.0"
rapidfuzz = ">=3.14.1,<4.0.0"
sqlite-utils = ">=3.38,<4.0"
prometheus-client = ">=0.23.1,<0.24.0"
python-telegram-bot = ">=22.5,<23.0"
apscheduler = ">=3.11.0,<4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.4.2,<9.0.0"
pytest-asyncio = ">=1.2.0,<2.0.0"
anyio = ">=4.11.0,<5.0.0"
ruff = ">=0.13.2,<0.14.0"
mypy = ">=1.18.2,<2.0.0"
types-python-dateutil = ">=2.9.0.20250822,<3.0.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
